Change back to seconds, uint32 for TimeStamp field

+++ backend/cen-v4.sql
@@ -4,7 +4,7 @@ DROP TABLE IF EXISTS TCNReport;
 CREATE TABLE `TCNReport` (
    `reportVK`	varchar(32) DEFAULT "",
    `report`     varchar(4000) DEFAULT "",
-   `reportTS` bigint unsigned,
+   `reportTS` int unsigned,
    PRIMARY KEY (`reportVK`,`reportTS`),
    KEY (`reportTS`)
 );
